Conflicts:
	interface/CMakeLists.txt
	interface/src/Application.cpp

+++ libraries/octree/src/OctreeElement.cpp
@@ -196,6 +196,7 @@ void OctreeElement::calculateAABox() {
 void OctreeElement::deleteChildAtIndex(int childIndex) {
     OctreeElement* childAt = getChildAtIndex(childIndex);
     if (childAt) {
+printf("deleteChildAtIndex()... about to call delete childAt=%p\n",childAt);
         delete childAt;
         setChildAtIndex(childIndex, NULL);
         _isDirty = true;
@@ -1119,24 +1120,33 @@ OctreeElement* OctreeElement::addChildAtIndex(int childIndex) {
 }
 
 // handles staging or deletion of all deep children
-void OctreeElement::safeDeepDeleteChildAtIndex(int childIndex, int recursionCount) {
+bool OctreeElement::safeDeepDeleteChildAtIndex(int childIndex, int recursionCount) {
+    bool deleteApproved = false;
     if (recursionCount > DANGEROUSLY_DEEP_RECURSION) {
         qDebug() << "OctreeElement::safeDeepDeleteChildAtIndex() reached DANGEROUSLY_DEEP_RECURSION, bailing!\n";
-        return;
+        return deleteApproved;
     }
     OctreeElement* childToDelete = getChildAtIndex(childIndex);
     if (childToDelete) {
-        // If the child is not a leaf, then call ourselves recursively on all the children
-        if (!childToDelete->isLeaf()) {
-            // delete all it's children
-            for (int i = 0; i < NUMBER_OF_CHILDREN; i++) {
-                childToDelete->safeDeepDeleteChildAtIndex(i,recursionCount+1);
+        if (childToDelete->deleteApproved()) {
+            // If the child is not a leaf, then call ourselves recursively on all the children
+            if (!childToDelete->isLeaf()) {
+                // delete all it's children
+                for (int i = 0; i < NUMBER_OF_CHILDREN; i++) {
+                    deleteApproved = childToDelete->safeDeepDeleteChildAtIndex(i,recursionCount+1);
+                    if (!deleteApproved) {
+                        break; // no point in continuing...
+                    }
+                }
+            }
+            if (deleteApproved) {
+                deleteChildAtIndex(childIndex);
+                _isDirty = true;
+                markWithChangedTime();
             }
         }
-        deleteChildAtIndex(childIndex);
-        _isDirty = true;
-        markWithChangedTime();
     }
+    return deleteApproved;
 }
 
 
@@ -1273,4 +1283,74 @@ void OctreeElement::notifyUpdateHooks() {
     for (int i = 0; i < _updateHooks.size(); i++) {
         _updateHooks[i]->elementUpdated(this);
     }
+}
+
+OctreeElement* OctreeElement::getOrCreateChildElementAt(float x, float y, float z, float s) {
+    OctreeElement* child = NULL;
+    // If the requested size is less than or equal to our scale, but greater than half our scale, then
+    // we are the Element they are looking for.
+    float ourScale = getScale();
+    float halfOurScale = ourScale / 2.0f;
+
+    assert(s <= ourScale); // This should never happen
+
+    if (s > halfOurScale) {
+        return this;
+    }
+    // otherwise, we need to find which of our children we should recurse
+    glm::vec3 ourCenter = _box.calcCenter();
+    
+    int childIndex = CHILD_UNKNOWN;
+    // left half
+    if (x > ourCenter.x) {
+        if (y > ourCenter.y) {
+            // top left
+            if (z > ourCenter.z) {
+                // top left far
+                childIndex = CHILD_TOP_LEFT_FAR;
+            } else {
+                // top left near
+                childIndex = CHILD_TOP_LEFT_NEAR;
+            }
+        } else {
+            // bottom left
+            if (z > ourCenter.z) {
+                // bottom left far
+                childIndex = CHILD_BOTTOM_LEFT_FAR;
+            } else {
+                // bottom left near
+                childIndex = CHILD_BOTTOM_LEFT_NEAR;
+            }
+        }
+    } else {
+        // right half
+        if (y > ourCenter.y) {
+            // top right
+            if (z > ourCenter.z) {
+                // top right far
+                childIndex = CHILD_TOP_RIGHT_FAR;
+            } else {
+                // top right near
+                childIndex = CHILD_TOP_RIGHT_NEAR;
+            }
+        } else {
+            // bottom right
+            if (z > ourCenter.z) {
+                // bottom right far
+                childIndex = CHILD_BOTTOM_RIGHT_FAR;
+            } else {
+                // bottom right near
+                childIndex = CHILD_BOTTOM_RIGHT_NEAR;
+            }
+        }
+    }
+    
+    // Now, check if we have a child at that location
+    child = getChildAtIndex(childIndex);
+    if (!child) {
+        child = addChildAtIndex(childIndex);
+    }
+    
+    // Now that we have the child to recurse down, let it answer the original question...
+    return child->getOrCreateChildElementAt(x, y, z, s);
 }
